refactor: extract shared HTTP client helpers
